site: reposition as Enterprise Architecture Reference Catalog
Reframe the site around three orthogonal artefacts (capabilities,
business processes, value streams) instead of capabilities only.

- New landing page at `/` with hero banner and three tiles linking to
  the three catalogues.
- Move the capability browser to `/capabilities`.
- Extract a generic `CatalogBrowser` (with `FilterBar`, exporters,
  detail modal) under `site/src/components/catalog/`. The legacy
  `CatalogueBrowser.tsx` becomes a thin wrapper for back-compat.
- `/business-processes` and `/value-streams` now use the same browser
  shell (search, level/industry filters, expand-by-level stepper,
  multi-select, CSV/JSON export, detail modal). Value streams are
  flattened into a stream→stage tree via a new adapter.
- Update Base layout title, nav order, and footer to match the new
  positioning; extend the icon font to include the new tile glyphs.
