+++ site/src/layouts/Base.astro
@@ -1,7 +1,9 @@
 ---
 import "../styles/global.css";
 import "../styles/catalogue.css";
-import { version } from "../data/load.ts";
+import { valueStreams, version } from "../data/load.ts";
+
+const streamCount = valueStreams.length;
 
 interface Props {
   title: string;
@@ -11,11 +13,11 @@ interface Props {
 
 const { title, description, navTransparent = false } = Astro.props;
 const pageTitle = title
-  ? `${title} · Business Capability Reference Catalog`
-  : "Business Capability Reference Catalog";
+  ? `${title} · Enterprise Architecture Reference Catalog`
+  : "Enterprise Architecture Reference Catalog";
 const pageDescription =
   description ??
-  "An open Business Capability Reference Catalog. Browse, search, and export a canonical, industry-aware capability hierarchy.";
+  "Open Enterprise Architecture Reference Catalog — capabilities, business processes, and value streams in one searchable, exportable model.";
 const navClass = navTransparent ? "nav nav-transparent" : "nav";
 const generatedDate = version.generated_at.slice(0, 10);
 ---
@@ -35,7 +37,7 @@ const generatedDate = version.generated_at.slice(0, 10);
     />
     <link
       rel="stylesheet"
-      href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@24,500,0,0&icon_names=add,arrow_upward,chevron_right,close,expand_more,remove"
+      href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@24,500,0,0&icon_names=account_tree,add,arrow_upward,chevron_right,close,expand_more,linear_scale,remove,route"
     />
     <link rel="icon" type="image/png" href="/favicon.png" />
     <meta property="og:title" content={pageTitle} />
@@ -55,9 +57,10 @@ const generatedDate = version.generated_at.slice(0, 10);
         </a>
         <div class="nav-links">
           <a href="/">Home</a>
-          <a href="/about">About</a>
-          <a href="/value-streams">Value Streams</a>
+          <a href="/capabilities">Capabilities</a>
           <a href="/business-processes">Processes</a>
+          <a href="/value-streams">Value Streams</a>
+          <a href="/about">About</a>
           <a href="/api/tree.json">API</a>
           <a
             href="https://www.turbo-ea.org"
@@ -96,7 +99,7 @@ const generatedDate = version.generated_at.slice(0, 10);
     <footer class="footer">
       <div class="footer-inner">
         <div class="footer-brand">
-          <span class="footer-brand-text">Business Capability</span>
+          <span class="footer-brand-text">Enterprise Architecture</span>
           <span class="footer-brand-sub">Reference Catalog</span>
         </div>
         <div class="footer-links">
@@ -122,7 +125,8 @@ const generatedDate = version.generated_at.slice(0, 10);
         </div>
         <div class="footer-copy">
           Catalog {version.catalogue_version} · schema v{version.schema_version} ·
-          {version.node_count} capabilities{version.process_count ? ` · ${version.process_count} processes` : ""} ·
+          {version.node_count} capabilities{version.process_count ? ` · ${version.process_count} processes` : ""}
+          {streamCount > 0 ? ` · ${streamCount} value streams` : ""} ·
           generated {generatedDate}
         </div>
       </div>
