+++ site/src/styles/catalogue.css
@@ -1316,3 +1316,151 @@
     height: 40px;
   }
 }
+
+/* ---------------------------------------------------------------------------
+ * Landing-page tile grid (three catalog tiles: capabilities / processes / VS)
+ * ------------------------------------------------------------------------- */
+.catalog-tile-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(280px, 1fr));
+  gap: 20px;
+  margin: 32px 0 24px;
+}
+
+.catalog-tile {
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+  padding: 24px;
+  background: var(--color-white);
+  border: 1px solid var(--color-border);
+  border-radius: var(--radius-sm, 8px);
+  text-decoration: none;
+  color: inherit;
+  transition: border-color 0.15s ease, box-shadow 0.15s ease, transform 0.15s ease;
+  position: relative;
+  min-height: 240px;
+}
+
+.catalog-tile:hover {
+  border-color: var(--color-secondary);
+  box-shadow: 0 6px 18px rgba(30, 58, 138, 0.12);
+  transform: translateY(-2px);
+}
+
+.catalog-tile:focus-visible {
+  outline: 2px solid var(--color-accent);
+  outline-offset: 2px;
+}
+
+.catalog-tile-icon {
+  width: 44px;
+  height: 44px;
+  border-radius: 10px;
+  background: var(--color-pale-blue);
+  color: var(--color-secondary);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.catalog-tile-icon .material-symbols-outlined {
+  font-size: 26px;
+}
+
+.catalog-tile-body {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  flex: 1;
+}
+
+.catalog-tile-title {
+  margin: 0;
+  font-size: 20px;
+  font-weight: 700;
+  color: var(--color-dark, #0f172a);
+}
+
+.catalog-tile-tagline {
+  margin: 0;
+  color: var(--color-muted);
+  font-size: 14px;
+  line-height: 1.5;
+}
+
+.catalog-tile-tagline strong {
+  color: var(--color-dark, #0f172a);
+  font-weight: 600;
+}
+
+.catalog-tile-count {
+  display: inline-flex;
+  align-items: baseline;
+  gap: 6px;
+  font-size: 22px;
+  font-weight: 700;
+  color: var(--color-secondary);
+  margin-top: auto;
+}
+
+.catalog-tile-count-label {
+  font-size: 12px;
+  font-weight: 500;
+  color: var(--color-muted);
+  text-transform: uppercase;
+  letter-spacing: 0.04em;
+}
+
+.catalog-tile-cta {
+  font-size: 13px;
+  font-weight: 600;
+  color: var(--color-accent);
+}
+
+.catalog-tile:hover .catalog-tile-cta {
+  text-decoration: underline;
+}
+
+.landing-meta {
+  margin: 24px 0 48px;
+  padding: 20px 24px;
+  background: var(--color-pale-blue);
+  border-radius: var(--radius-sm, 8px);
+  color: var(--color-muted);
+  font-size: 14px;
+  line-height: 1.6;
+}
+
+.landing-meta p {
+  margin: 0;
+}
+
+.landing-meta p + p {
+  margin-top: 8px;
+}
+
+.landing-meta code {
+  background: var(--color-white);
+  padding: 1px 5px;
+  border-radius: 3px;
+  font-size: 12px;
+  color: var(--color-secondary);
+}
+
+.landing-meta-links {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 10px;
+  align-items: center;
+}
+
+.landing-meta-links a {
+  color: var(--color-accent);
+  text-decoration: none;
+  font-weight: 500;
+}
+
+.landing-meta-links a:hover {
+  text-decoration: underline;
+}
